+++ src/components/amdin/Category/DeleteSecondPopup.vue
@@ -19,6 +19,8 @@
 import { defineProps, defineEmits, ref } from 'vue';
 import { useStore } from 'vuex';
 import CategoryList from "@/components/amdin/Category/CategoryList.vue";
+import Swal from "sweetalert2";
+
 const store = useStore();
 const accessToken = store.state.accessToken;
 
@@ -40,7 +42,25 @@ const closeDeleteModal = () => {
 const deleteCategorySecond = async () => {
 
   try {
-    const response = await fetch(`http://localhost:5000/admin/category/second/delete/${props.currentSecondCode}`, {
+    const categoryThirdResponse = await fetch(`http://api.pioms.shop/admin/category/third/list/detail/categorysecond/${props.currentSecondCode}`, {
+      method: 'GET',
+      headers: {
+        'Authorization': `Bearer ${accessToken}`,
+        'Content-Type': 'application/json',
+      },
+    })
+    const categoryThirds = await categoryThirdResponse.json();
+
+    if(categoryThirds.length > 0) {
+      await Swal.fire({
+        icon: 'success',
+        title: '중분류 카테고리 삭제 실패',
+        text: '카테고리를 삭제할 수 없습니다. 하위 카테고리가 존재합니다.',
+      });
+      return;
+    }
+
+    const response = await fetch(`http://api.pioms.shop/admin/category/second/delete/${props.currentSecondCode}`, {
       method: 'DELETE',
       headers: {
         'Authorization': `Bearer ${accessToken}`,
@@ -48,7 +68,11 @@ const deleteCategorySecond = async () => {
       },
     });
 
-    console.log('카테고리 대분류가 성공적으로 수정되었습니다.');
+    await Swal.fire({
+      icon: 'success',
+      title: '카테고리 삭제 성공!',
+      text: '중분류 카테고리를 삭제하였습니다.',
+    });
     location.reload(CategoryList);
     emits('close');
   } catch (error) {
@@ -115,11 +139,13 @@ const closePopup = () => {
 
 .confirm-button {
   background-color: #344DAF;
+  border: none;
   color: white;
 }
 
 .cancel-button {
   background-color: #FF6285;
+  border: none;
   color: white;
 }
 </style>
